// src/types/enums/typeEnum.ts
export enum typeEnum {
  income = 'income',
  expense = 'expense',
}

export enum typeEnumsAl {
  income = 'Hyrje',
  expense = 'Dalje',
}

export enum currencyEnums {
  euro = 'euro',
  dinar = 'dinar',
  dollar = 'dollar',
  frang = 'frang',
  other = 'tjeter',
}

export enum actionsEnum {
  add = 'add',
  edit = 'edit',
  emptyEdit = 'emptyEdit',
}